+++ src/index.ts
@@ -6,4 +6,4 @@ export type {SystemType} from './system';
 export {component, field, system} from './decorators';
 export {Query} from './query';
 export type {ComponentType} from './component';
-export type {Frame, SystemGroup} from './schedules';
+export type {Frame, SystemGroup} from './schedule';