+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Upcoming
 - Allow duplicate component and system types in world defs.  They'll be deduplicated automatically, though only one copy of a system type is allowed to have initial props specified.
+- Record some basic stats about systems in `World.stats`.
+- Accept a scheduling definition function in the `@system` decorator.  This is equivalent to but requires less boilerplate than invoking `this.schedule` in the system itself.
+- Fix `@system` decorator not adding system to world defs if it was passed an argument.
 
 ### 0.9.1
 - Fix crash when using backref fields in components with an elastic storage type.